requirements.txt
Django==4.2.7
djangorestframework==3.15.1
drf-spectacular==0.27.1
python-dotenv==1.0.1
channels==4.0.0